Restructured damage system
Weapons should always be Area2D with knockback
applied through scripts.
This is to avoid breaking the physics engine with usually
incredibly fast weapons.

Damage-dealers search for victims, while victims announce themselves.
That is, if something is to be damaged with a sword,
it should be in the `sword_damageable` group.
In turn, when the sword collides with something,
it checks if it's in said group before dealing damage.

One quirk to note is that it seems Area2D can detect both bodies
and other Area2Ds, but bodies can only detect other bodies.

In turn, Area2D, if the only Area2D in a collision,
should be the "main actor" of said collision;
it should recieve and process the collision info.

https://godotengine.org/qa/16577/how-to-detect-a-collision-with-an-area2d